+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 
 - Fix returns values depending on the language
 - Fix display edit meta data for attribute availability has not i18n en_US
+- Fix slug length to 50 characters